+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 - muted_noteテーブルは使われなくなったため手動で削除を行ってください。
 
 ### Changes
-- API: users/notes で fileType 指定はできなくなりました
+- API: users/notes, notes/local-timeline で fileType 指定はできなくなりました
 
 ### Server
 - タイムライン取得時のパフォーマンスを改善